chore: regroup flatten functions
